+++ infra/report/rp_model_impl/json.go
@@ -126,7 +126,7 @@ func (z *Json) Close() {
 	if z.file != nil {
 		z.file.Close()
 
-		if z.index < 1 {
+		if z.index < 1 && z.ctl.IsProduction() {
 			z.ctl.Log().Debug("Try removing empty report file", zap.String("path", z.path))
 			err := os.Remove(z.path)
 			z.ctl.Log().Debug("Removed or had an error (ignore)", zap.String("path", z.path), zap.Error(err))
